penheal/requirement.txt
openai
requests
docker
scapy
python-dotenv
transformers
